TD-3897-DataService reference removed from the controller class. Service class added. changes made in respective classes.
